moves card component into its own folder
