+++ scripts/components/card/card.component.js
@@ -1,4 +1,5 @@
-import createEle from "../utils/createEle";
+import './card.styles.scss'  
+import createEle from "../../utils/createEle";
 
 const renderDinoCard = (dino, callToActions, randomCallToAction, i) => {
     const renderDinoPopup = (dino) =>
